+++ src/cosmos/info/thorchainruneInfo.ts
@@ -1,10 +1,21 @@
-import { EdgeCurrencyInfo } from 'edge-core-js/types'
+import { EdgeCurrencyInfo, EdgeTokenMap } from 'edge-core-js/types'
 
 import { makeOuterPlugin } from '../../common/innerPlugin'
 import type { CosmosTools } from '../CosmosTools'
 import { asCosmosInfoPayload, CosmosInfoPayload } from '../cosmosTypes'
 import { ThorchainNetworkInfo } from '../thorchainTypes'
 // import { cosmosCustomTokenTemplate } from './cosmosCommonInfo'
+
+const builtinTokens: EdgeTokenMap = {
+  tcy: {
+    currencyCode: 'TCY',
+    displayName: 'TCY',
+    denominations: [{ name: 'TCY', multiplier: '100000000' }],
+    networkLocation: {
+      contractAddress: 'tcy'
+    }
+  }
+}
 
 const networkInfo: ThorchainNetworkInfo = {
   bech32AddressPrefix: 'thor',
@@ -86,6 +97,7 @@ export const thorchainrune = makeOuterPlugin<
   currencyInfo,
   asInfoPayload: asCosmosInfoPayload,
   networkInfo,
+  builtinTokens,
 
   checkEnvironment() {
     if (global.BigInt == null) {